jewel sword working across versions
