+++ config/splat.pspeu.strno3.yaml
@@ -66,7 +66,7 @@ segments:
       - [0x18860, c, rno3/st_common]
       - [0x19FB8, c, rno3/prim_helpers]
       - [0x1A748, c, rno3_psp/unk_1A748]
-      - [0x1C7D8, c, rno3_psp/e_jewel_sword_puzzle]
+      - [0x1C7D8, c, rno3/e_jewel_sword_puzzle]
       - [0x1D498, c, rno3_psp/e_stairway]
       - [0x1DD38, c, rno3/e_background_bushes_trees]
       - [0x1E640, c, rno3_psp/st_init]
@@ -82,7 +82,7 @@ segments:
       - [0x212E8, c, rno3/e_sky_entities]
       - [0x214E0, c, rno3_psp/func_us_801b3c88]
       - [0x21780, data]
-      - [0x23138, .data, rno3_psp/e_jewel_sword_puzzle]
+      - [0x23138, .data, rno3/e_jewel_sword_puzzle]
       - [0x23248, .data, rno3/e_background_bushes_trees]
       - [0x232F8, data]
       - [0x2CAC8, .data, rno3/e_init]
