+++ config/splat.us.strno3.yaml
@@ -53,7 +53,8 @@ segments:
       - [0xF84, .data, e_background_lightning]
       - [0xF9C, .data, e_sky_entities]
       - [0xFB4, .data, e_cavern_door]
-      - [0xFCC, data]
+      - [0xFC0, .data, e_jewel_sword_puzzle]
+      - [0x10C0, data]
       - [0x1550, .data, st_common]
       - [0x1750, data]
       - [0x2FF60, .rodata, e_castle_door]
